reorganizing amatica transfers, starting rstar prep
